setting up backend-api call
