+++ .gitignore
@@ -60,3 +60,5 @@ typings/
 .node_modules
 .openode
 # config/
+
+.vscode
