+++ app/src/components/blogs/Blogs.jsx
@@ -9,17 +9,7 @@ import styles from "./styles.module.css";
 import { StickyContainer, Sticky } from "react-sticky"
 import { FaUserAlt, FaCommentDots} from "react-icons/fa";
 import { AiFillHeart, AiFillCalendar, AiFillEye} from "react-icons/ai";
-import kashi from './data/Kashi-Vishwanath-Darshan.json'
-import places from './data/Top-10-Places-in-Varanasi.json'
-import rental from './data/carHire.json'
-import Ghats from './data/Ghats.json'
-import Bhu from './data/Bhu.json'
-import Devd from "./data/Dev-dipawali.json"
-import Prayagraj from "./data/Prayagraj.json"
-import Vindhyachal from "./data/Vindhyachal.json"
-import Gaya from "./data/Gaya.json"
-import Sarnath from "./data/Sarnath.json"
-import Durga from "./data/Durgamandir.json"
+
 
 function Blogs({match}) {
   const route = match.params.name;
@@ -28,16 +18,13 @@ function Blogs({match}) {
   
   const dataAbout = async () => {
     try {
-      var url = `/${route}`;
-      var request = {
-        url,
-        method: "get",
-      };
-      const res = await axios(route);
-      const result = await res.data;
-      setData(result);
+      
+      const res = await axios.get(`http://localhost:7071/api/getBlogs?name=${route}`);
+      const result =  res;
+      console.log(result.data)
+      setData(result.data);
       localStorage.setItem("data", JSON.stringify(result))
-    } catch (err) {
+    } catch(err) {
       let collection = localStorage.getItem("data");
       setData(JSON.parse(collection))
       console.log(err);
@@ -47,28 +34,7 @@ function Blogs({match}) {
   useEffect(() => {
     dataAbout();
   }, [route]);
-  if(route ==="Kashi Vishwanath")
-  data = kashi;
-  else if(route === "carRental")
-  data = rental 
-  else if(route === "Ghats")
-  data = Ghats
-  else if(route === "Bhu")
-  data = Bhu
-  else if(route === "DevDipawali")
-  data = Devd
-  else if(route === "Top 10 places")
-  data = places;
-  else if(route === "Prayagraj")
-  data = Prayagraj;
-  else if(route === "Vindhyachal")
-  data = Vindhyachal;
-  else if(route === "Gaya")
-  data = Gaya;
-  else if(route === "Sarnath Darshan")
-  data = Sarnath;
-  else if(route === "Durga Mandir")
-  data = Durga;
+  
 
   const { title, intro, article, keyword, description, pageName, image } = data;
   return (
