+++ app/src/components/navbar/NavbarComponent.jsx
@@ -68,17 +68,17 @@ function NavbarComponent() {
                 <Link to="/blogs/Gaya" className="font-15 text-brown font-regular dropdown-item">
                   Gaya
                 </Link>
-                <Link to="/blogs/Sarnath Darshan" className="font-15 text-brown font-regular dropdown-item">
+                <Link to="/blogs/Sarnath" className="font-15 text-brown font-regular dropdown-item">
                   Sarnath
                 </Link>
-                <Link to="/blogs/Durga Mandir" className="font-15 text-brown font-regular dropdown-item">
+                <Link to="/blogs/Durgamandir" className="font-15 text-brown font-regular dropdown-item">
                   Durga Mata Mandir
                 </Link>
               </div>
             </NavDropdown>
             <NavDropdown title="Packages" id="collasible-nav-dropdown" className="font-medium font-17 px-3 navbar-item text-white text-center active">
               <div style={{background: '#FFECD3', border: 'none', borderRadius: '5px', color: '#663D14'}}>
-                <Link to="/package/airport" className="font-15 text-brown font-regular dropdown-item">                  
+                <Link to="/package/Top 10 places" className="font-15 text-brown font-regular dropdown-item">                  
                   Airport                
                 </Link>
                 <Link to="/package/Vindhyachal" className="font-15 text-brown font-regular dropdown-item">                  
@@ -92,7 +92,7 @@ function NavbarComponent() {
                 </Link>
               </div>
             </NavDropdown>
-            <NavLink to="/blogs/carRental" activeClassName="activeNav">
+            <NavLink to="/blogs/carHire" activeClassName="activeNav">
               <Nav.Link
                 href="/blogs/carRental"
                 className="font-medium font-17 px-3 navbar-item
