+++ app/src/swa-cli.config.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://aka.ms/azure/static-web-apps-cli/schema",
+  "configurations": {
+    "src": {
+      "appLocation": ".",
+      "outputLocation": "."
+    }
+  }
+}
